requirements.txt
langchain
python-dotenv
streamlit
unstructured
tiktoken
faiss-cpu
# libmagic==1.0
# python-magic==0.4.27
# python-magic-bin==0.4.14
langchain_cohere